Fixed Tamper integration.
The form logic for the derived Tamper process plugin is moved to a MigrateForm plugin called 'TamperForm'. This class designates all form related calls to the Tamper plugin. Also added is that source definitions are read from the Migrate entity. This ensures that the Tamper Rewrite plugin can display replacement patterns.